updates path
