+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "frontend-interview-bot",
-  "version": "1.0.0",
+  "version": "1.1.0",
   "description": "Itnerview-bot-JavaScript",
   "main": "index.js",
   "scripts": {
